+++ t21log/src/test/java/com/tempos21/android/commons/utils/T21FileLogTest.java
@@ -18,23 +18,30 @@ import java.io.IOException;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
+import static org.mockito.Matchers.anyString;
+import static org.mockito.Matchers.eq;
+import static org.mockito.Mockito.never;
+import static org.mockito.Mockito.times;
+import static org.powermock.api.mockito.PowerMockito.verifyStatic;
 
 @RunWith(PowerMockRunner.class)
 @PrepareForTest({Log.class})
-public class FileLoggerTest {
+public class T21FileLogTest {
+
+    private static final String TAG = "[T21FileLogTest]";
 
     private static File logFile;
 
-    private static final String DATE_TIME_FORMAT = "yyyy/MM/dd HH:mm:ss.SSS";
-
     private static final String DATE_TIME_REGEX = "\\d{4}/\\d{2}/\\d{2} \\d{2}:\\d{2}:\\d{2}\\.\\d{3}";
 
-    private static String TAG = "[FileLoggerTest]";
+    private static final String HOLA = "hola";
+
+    private static final String DOUBLE_COLON = "::";
 
     @BeforeClass
     public static void setUpOnce() {
-        logFile = new File("FileLoggerTest.log");
-        FileLogger.initialize(logFile);
+        logFile = new File("T21FileLogTest.log");
+        T21FileLog.initialize(TAG, true, logFile);
     }
 
     @Before
@@ -45,82 +52,127 @@ public class FileLoggerTest {
     @Test
     public void verbose() {
         String VERBOSE_TAG = "V";
+        int i = 452;
         BuildingTest buildingTest = new BuildingTest();
         buildingTest.setId(2);
         buildingTest.setName("nameTestVerbose");
         buildingTest.setAddress("addressTestVerbose");
 
-        FileLogger.v(TAG, buildingTest.toString());
+        T21FileLog.v(HOLA, i, DOUBLE_COLON, buildingTest);
 
         assertTrue("Log tag '" + TAG + "' not found in log file", readFileLastLine().contains(TAG));
         assertEquals("Verbose tag '" + VERBOSE_TAG + "' not found in log file", VERBOSE_TAG, getLogLevelTag());
         assertTrue("Date and time not found in log file", getTimestamp().matches(DATE_TIME_REGEX));
-        assertEquals("Object logged '" + buildingTest + "' not found in log file", buildingTest.toString(), getMessage());
+        assertEquals("Object logged '" + buildingTest + "' not found in log file",
+                HOLA + " " + i + " " + DOUBLE_COLON + " " + buildingTest, getMessage());
     }
 
     @Test
     public void debug() {
         String DEBUG_TAG = "D";
+        int i = 71;
         BuildingTest buildingTest = new BuildingTest();
         buildingTest.setId(0);
         buildingTest.setName("nameTestDebug");
         buildingTest.setAddress("addressTestDebug");
 
-        FileLogger.d(TAG, buildingTest.toString());
+        T21FileLog.d(HOLA, i, DOUBLE_COLON, buildingTest);
 
         assertTrue("Log tag '" + TAG + "' not found in log file", readFileLastLine().contains(TAG));
         assertEquals("Debug tag '" + DEBUG_TAG + "' not found in log file", DEBUG_TAG, getLogLevelTag());
         assertTrue("Date and time not found in log file", getTimestamp().matches(DATE_TIME_REGEX));
-        assertEquals("Object logged '" + buildingTest + "' not found in log file", buildingTest.toString(), getMessage());
+        assertEquals("Object logged '" + buildingTest + "' not found in log file",
+                HOLA + " " + i + " " + DOUBLE_COLON + " " + buildingTest, getMessage());
     }
 
     @Test
     public void info() {
         String INFO_TAG = "I";
+        int i = 567;
         BuildingTest buildingTest = new BuildingTest();
         buildingTest.setId(3);
         buildingTest.setName("nameTestInfo");
         buildingTest.setAddress("addressTestInfo");
 
-        FileLogger.i(TAG, buildingTest.toString());
+        T21FileLog.i(HOLA, i, DOUBLE_COLON, buildingTest);
 
         assertTrue("Log tag '" + TAG + "' not found in log file", readFileLastLine().contains(TAG));
         assertEquals("Info tag '" + INFO_TAG + "' not found in log file", INFO_TAG, getLogLevelTag());
         assertTrue("Date and time not found in log file", getTimestamp().matches(DATE_TIME_REGEX));
-        assertEquals("Object logged '" + buildingTest + "' not found in log file", buildingTest.toString(), getMessage());
+        assertEquals("Object logged '" + buildingTest + "' not found in log file",
+                HOLA + " " + i + " " + DOUBLE_COLON + " " + buildingTest, getMessage());
     }
 
     @Test
     public void warning() {
         String WARNING_TAG = "W";
+        int i = 81;
         BuildingTest buildingTest = new BuildingTest();
         buildingTest.setId(12);
         buildingTest.setName("nameTestWarning");
         buildingTest.setAddress("addressTestWarning");
 
-        FileLogger.w(TAG, buildingTest.toString());
+        T21FileLog.w(HOLA, i, DOUBLE_COLON, buildingTest);
 
         assertTrue("Log tag '" + TAG + "' not found in log file", readFileLastLine().contains(TAG));
         assertEquals("Warning tag '" + WARNING_TAG + "' not found in log file", WARNING_TAG, getLogLevelTag());
         assertTrue("Date and time not found in log file", getTimestamp().matches(DATE_TIME_REGEX));
-        assertEquals("Object logged '" + buildingTest + "' not found in log file", buildingTest.toString(), getMessage());
+        assertEquals("Object logged '" + buildingTest + "' not found in log file",
+                HOLA + " " + i + " " + DOUBLE_COLON + " " + buildingTest, getMessage());
     }
 
     @Test
     public void error() {
         String ERROR_TAG = "E";
-
+        int i = 78;
         BuildingTest buildingTest = new BuildingTest();
         buildingTest.setId(99);
         buildingTest.setName("nameTestError");
         buildingTest.setAddress("addressTestError");
 
-        FileLogger.e(TAG, buildingTest.toString());
+        T21FileLog.e(HOLA, i, DOUBLE_COLON, buildingTest);
 
         assertTrue("Log tag '" + TAG + "' not found in log file", readFileLastLine().contains(TAG));
         assertEquals("Error tag '" + ERROR_TAG + "' not found in log file", ERROR_TAG, getLogLevelTag());
         assertTrue("Date and time not found in log file", getTimestamp().matches(DATE_TIME_REGEX));
-        assertEquals("Object logged '" + buildingTest + "' not found in log file", buildingTest.toString(), getMessage());
+        assertEquals("Object logged '" + buildingTest + "' not found in log file",
+                HOLA + " " + i + " " + DOUBLE_COLON + " " + buildingTest, getMessage());
+    }
+
+    @Test
+    public void writeToFile() {
+        File file = new File("T21LogTest_writeToFile.log");
+        T21FileLog.initialize(TAG, true, file);
+        int i = 34;
+        BuildingTest buildingTest = new BuildingTest();
+        buildingTest.setId(28);
+        buildingTest.setName("nameTestFile");
+        buildingTest.setAddress("addressTestFile");
+
+        T21Log.i(HOLA, i, DOUBLE_COLON, buildingTest);
+
+        verifyStatic(times(1));
+        T21FileLog.i(eq(TAG), anyString());
+
+        file.delete();
+    }
+
+    @Test
+    public void disabledLogWithFile() {
+        File file = new File("T21LogTest_disabledLogWithFile.log");
+        T21FileLog.initialize(TAG, false, file);
+        int i = 87;
+        BuildingTest buildingTest = new BuildingTest();
+        buildingTest.setId(3698);
+        buildingTest.setName("nameTestDisabledWithFile");
+        buildingTest.setAddress("addressTestDisabledWithFile");
+
+        T21FileLog.v(HOLA, i, DOUBLE_COLON, buildingTest);
+
+        verifyStatic(never());
+        T21FileLog.v(eq(TAG), anyString());
+
+        file.delete();
     }
 
     private static String getLogLevelTag() {
